Moved ug components into new intl/ug folder
